Fix period operation type autoloading and rule compatibility
