+++ orchestrators/SettingsManagement/src/DTOs/FiscalPeriod/PeriodOperationType.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Nexus\SettingsManagement\DTOs\FiscalPeriod;
+
+/**
+ * Operation types for period validation.
+ */
+enum PeriodOperationType: string
+{
+    case TRANSACTION = 'transaction';
+    case ADJUSTMENT = 'adjustment';
+    case CLOSING = 'closing';
+    case REPORTING = 'reporting';
+}
