+++ orchestrators/SettingsManagement/src/DTOs/FiscalPeriod/PeriodValidationRequest.php
@@ -5,17 +5,6 @@ declare(strict_types=1);
 namespace Nexus\SettingsManagement\DTOs\FiscalPeriod;
 
 /**
- * Operation types for period validation.
- */
-enum PeriodOperationType: string
-{
-    case TRANSACTION = 'transaction';
-    case ADJUSTMENT = 'adjustment';
-    case CLOSING = 'closing';
-    case REPORTING = 'reporting';
-}
-
-/**
  * Request DTO for validating fiscal period status.
  */
 final class PeriodValidationRequest
